backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "authlib>=1.6.9",
    "fastapi[standard]>=0.135.1",
    "httpx>=0.28.1",
    "itsdangerous>=2.2.0",
    "langchain-openai>=1.1.12",
    "langchain-pinecone>=0.2.13",
    "langgraph>=1.1.3",
    "litellm>=1.82.6",
    "pinecone-client>=6.0.0",
    "psycopg2-binary>=2.9.11",
    "pyjwt>=2.12.0",
    "pypdf>=6.9.2",
    "python-multipart>=0.0.22",
    "sqlalchemy[asyncio]>=2.0.48",
    "supabase>=2.28.3",
    "tavily-python>=0.7.23",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "ruff>=0.15.5",
]

[tool.fastapi]
entrypoint = "app.main:app"
